Only use the define replacement if necessary
We have a problem when using third party libraries that import things using AMD, but don't provide the file themselves.

For example,

```
if (typeof define === "function" && define.amd) {
	// AMD. Register as an anonymous module.
	define(["jquery"], factory);
} else {
	// Browser globals
	factory(jQuery);
}
```

In this case, the define replacement was used (`typeof define === "function"` was matched, and `define.amd` was matched.

To prevent this, only use the define replacement if there are recognised imports.